# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MK Store Backend API",
    "version": "1.0.0",
    "description": "E-commerce application backend API documentation with sample requests and responses",
    "contact": {
      "name": "API Support",
      "email": "support@mkstore.com"
    }
  },
  "servers": [
    {
      "url": "",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token obtained from login endpoints"
      }
    },
...
  },
...
  "paths": {
...
    "/api/save-product-variant": {
      "post": {
        "summary": "Create a new product variant",
        "tags": [
          "Product Variants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "productId",
                  "variantName",
                  "price",
                  "sellingPrice",
                  "quantity"
                ],
                "properties": {
                  "productId": {
                    "type": "integer",
                    "example": 1
                  },
                  "variantName": {
                    "type": "string",
                    "example": "500g"
                  },
                  "description": {
                    "type": "string",
                    "example": "Fresh product description",
                    "description": "Variant description (optional)"
                  },
                  "nutritional": {
                    "type": "string",
                    "example": "Calories: 100, Protein: 5g",
                    "description": "Nutritional information (optional)"
                  },
                  "price": {
                    "type": "number",
                    "example": 500
                  },
                  "sellingPrice": {
                    "type": "number",
                    "example": 450
                  },
                  "quantity": {
                    "type": "integer",
                    "example": 100
                  },
                  "itemQuantity": {
                    "type": "number",
                    "example": 0.5
                  },
                  "itemUnit": {
                    "type": "string",
                    "enum": [
                      "KG",
                      "G",
                      "LTR",
                      "ML",
                      "etc."
                    ],
                    "example": "KG"
                  },
                  "expiryDate": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-12-31",
                    "description": "Variant expiry date (required)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product variant created successfully"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/get-product-variants": {
      "post": {
        "summary": "Get product variants with pagination and filtering",
        "tags": [
          "Product Variants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "productId"
                ],
                "properties": {
                  "productId": {
                    "type": "integer",
                    "example": 1
                  },
                  "pageSize": {
                    "type": "integer",
                    "enum": [
                      1,
                      5,
                      10,
                      20,
                      30,
                      40,
                      50,
                      100,
                      500
                    ],
                    "default": 10
                  },
                  "pageNumber": {
                    "type": "integer",
                    "minimum": 1,
                    "default": 1
                  },
                  "filters": {
                    "type": "array"
                  },
                  "sorting": {
                    "type": "array"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product variants retrieved successfully"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/get-variant-by-id/{variantId}": {
      "get": {
        "summary": "Get product variant by ID",
        "tags": [
          "Product Variants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "variantId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product variant retrieved successfully"
          },
          "404": {
            "description": "Variant not found"
          }
        }
      }
    },
    "/api/update-product-variant/{id}": {
      "post": {
        "summary": "Update product variant",
        "tags": [
          "Product Variants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "concurrencyStamp"
                ],
                "properties": {
                  "variantName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string",
                    "description": "Variant description (optional)"
                  },
                  "nutritional": {
                    "type": "string",
                    "description": "Nutritional information (optional)"
                  },
                  "price": {
                    "type": "number"
                  },
                  "sellingPrice": {
                    "type": "number"
                  },
                  "quantity": {
                    "type": "integer"
                  },
                  "expiryDate": {
                    "type": "string",
                    "format": "date",
                    "example": "2024-12-31",
                    "description": "Variant expiry date (optional)"
                  },
                  "concurrencyStamp": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product variant updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Concurrency error"
          }
        }
      }
    },
    "/api/delete-product-variant/{id}": {
      "post": {
        "summary": "Delete product variant (soft delete)",
        "tags": [
          "Product Variants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product variant deleted successfully"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/get-variants-by-type/{productId}": {
      "get": {
        "summary": "Get product variants grouped by type",
        "tags": [
          "Product Variants"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Variants grouped by type"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
  }
}